requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
filelock==3.20.0
pillow==11.3.0
opencv-python-headless==4.8.0.76
jinja2==3.1.6
email-validator==2.0.0
schedule==1.2.0
cachetools==5.4.0
python-dotenv==1.0.0
playwright==1.40.0
tensorflow==2.17.0
numpy==1.26.4
scipy==1.11.4
h5py==3.10.0
imageio==2.31.1
matplotlib==3.8.2
gradio==5.49.1
orjson==3.11.3
gunicorn==21.2.0
torch==2.4.0
torchvision==0.19.0
safetensors==0.4.5
requests==2.32.5
reportlab==4.0.0
huggingface_hub==0.36.0
pydantic==2.11.10
pydantic-core==2.33.2
annotated-types==0.7.0
typing-inspection==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
rich==14.2.0
markdown-it-py==4.0.0
mdurl==0.1.2
APScheduler==3.10.4
itsdangerous==2.1.2
fastapi-sessions[redis]==0.3.2
